Remove redundant copy operations in FactorHiGHSSolver Add Zed editor caches to .gitignore Fix a compiler warning about potential infinite loops

+++ highs/ipm/hipo/ipm/Solver.cpp
@@ -2,6 +2,7 @@
 
 #include <cassert>
 #include <cmath>
+#include <cstdint>
 #include <iostream>
 
 #include "ipm/hipo/auxiliary/AutoDetect.h"
@@ -417,7 +418,7 @@ double Solver::stepToBoundary(const std::vector<double>& x,
   double alpha = 1.0;
   Int bl = -1;
 
-  for (Int i = 0; i < x.size(); ++i) {
+  for (uint64_t i = 0; i < x.size(); ++i) {
     if ((lo && model_.hasLb(i)) || (!lo && model_.hasUb(i))) {
       double c = (cor ? (*cor)[i] * weight : 0.0);
       if (x[i] + alpha * (dx[i] + c) < 0.0) {
